Validate bag info at load time
If we can't even read the bag info (e.g. because it's the wrong class) we'll fail before this validation. But in case we get the right class (or at least something readable), we further check all the properties. This could be extended to a more lenient version check in the future, and the actual checking function is broken out so that anyone subclassing this can do so freely.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>