Reintroduce fudge factor for context
I was just getting lucky initially that pinning the env variables was making the CI strictly faster here. More runs have shown it is sometimes a little slower. (Again, only on the CI, locally where there is not the same load interference, it works reliably). The fudge-factor then activates conditionally when we're running on GitHub.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>