Update benchmark criteria (#69)
* Relax BIC improvement threshold

The TrieH5Bag was sometimes erroneously better scaling than expected. It is a tricky one since the trie itself varies between linear and quadratic scaling.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Reintroduce fudge factor for context

I was just getting lucky initially that pinning the env variables was making the CI strictly faster here. More runs have shown it is sometimes a little slower. (Again, only on the CI, locally where there is not the same load interference, it works reliably). The fudge-factor then activates conditionally when we're running on GitHub.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Relax BIC improvement threshold

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Increase BIC threshold

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Increase BIC threshold

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

---------

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>